intelliShare/requirements.txt
flask
flask-bootstrap
flask-sqlalchemy